requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.2.1
numpy==1.26.2
pandas==2.1.4
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
